requirements.txt
# E2B Data Apps Builder - Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=12.0
python-multipart>=0.0.9

# AI Agent
claude-agent-sdk>=0.1.0

# E2B Sandbox
e2b-code-interpreter>=1.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
